server/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
sqlalchemy = "^2.0.27"
python-dotenv = "^1.0.1"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.2.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
